Add a way to run a callback on file load
This is useful when integrating syn-inline-mod with other systems, as it allows
for file contents to be recorded for e.g. later error reporting.

This required some refactoring as we had to pass a `&mut R` down through the
visitor.

Included is a test which ensures that syn-inline-mod works on the crate itself.